fix(compose): Move it out of the exporters directory
